+++ application/application-standard/src/test/java/com/milesight/beaveriot/sample/rule/interceptor/DemoRuleNodeInterceptor.java
@@ -1,9 +1,9 @@
 package com.milesight.beaveriot.sample.rule.interceptor;
 
 import com.milesight.beaveriot.rule.RuleNodeInterceptor;
-import com.milesight.beaveriot.rule.model.flow.route.FromNode;
-import com.milesight.beaveriot.rule.model.flow.route.RouteNode;
-import com.milesight.beaveriot.rule.model.flow.route.base.OutputNode;
+import com.milesight.beaveriot.rule.model.flow.yaml.FromNode;
+import com.milesight.beaveriot.rule.model.flow.yaml.RouteNode;
+import com.milesight.beaveriot.rule.model.flow.yaml.base.OutputNode;
 
 /**
  * @author leon
