+++ core/rule-engine/rule-engine-api/src/main/java/com/milesight/beaveriot/rule/model/flow/route/ToNodeDefinition.java
@@ -1,4 +1,4 @@
-package com.milesight.beaveriot.rule.model.flow.node;
+package com.milesight.beaveriot.rule.model.flow.route;
 
 import com.milesight.beaveriot.rule.model.definition.ComponentDefinition;
 import com.milesight.beaveriot.rule.model.flow.config.RuleNodeConfig;
