+++ core/rule-engine/rule-engine-api/src/main/java/com/milesight/beaveriot/rule/model/flow/route/WhenNodeDefinition.java
@@ -1,4 +1,4 @@
-package com.milesight.beaveriot.rule.model.flow.node;
+package com.milesight.beaveriot.rule.model.flow.route;
 
 import com.milesight.beaveriot.rule.model.flow.ExpressionNode;
 import lombok.Data;
