+++ core/rule-engine/rule-engine-api/src/main/java/com/milesight/beaveriot/rule/model/flow/yaml/base/MultiOutputNode.java
@@ -1,4 +1,4 @@
-package com.milesight.beaveriot.rule.model.flow.route.base;
+package com.milesight.beaveriot.rule.model.flow.yaml.base;
 
 import java.util.List;
