+++ core/rule-engine/rule-engine-api/src/main/java/com/milesight/beaveriot/rule/model/flow/yaml/base/NodeId.java
@@ -1,4 +1,4 @@
-package com.milesight.beaveriot.rule.model.flow.route.base;
+package com.milesight.beaveriot.rule.model.flow.yaml.base;
 
 /**
  * @author leon
