+++ core/rule-engine/rule-engine-core/src/main/java/com/milesight/beaveriot/rule/flow/yaml/DefaultRuleNodeInterceptor.java
@@ -1,9 +1,9 @@
-package com.milesight.beaveriot.rule.flow.builder;
+package com.milesight.beaveriot.rule.flow.yaml;
 
 import com.milesight.beaveriot.rule.RuleNodeInterceptor;
-import com.milesight.beaveriot.rule.model.flow.route.FromNode;
-import com.milesight.beaveriot.rule.model.flow.route.RouteNode;
-import com.milesight.beaveriot.rule.model.flow.route.base.OutputNode;
+import com.milesight.beaveriot.rule.model.flow.yaml.FromNode;
+import com.milesight.beaveriot.rule.model.flow.yaml.RouteNode;
+import com.milesight.beaveriot.rule.model.flow.yaml.base.OutputNode;
 
 import java.util.ArrayList;
 import java.util.Comparator;
