+++ core/rule-engine/rule-engine-core/src/main/java/com/milesight/beaveriot/rule/flow/yaml/RuleFlowGraph.java
@@ -1,4 +1,4 @@
-package com.milesight.beaveriot.rule.flow.builder;
+package com.milesight.beaveriot.rule.flow.yaml;
 
 import com.google.common.graph.GraphBuilder;
 import com.google.common.graph.MutableGraph;
