+++ core/rule-engine/rule-engine-core/src/main/java/com/milesight/beaveriot/rule/flow/yaml/RuleFlowYamlBuilder.java
@@ -1,4 +1,4 @@
-package com.milesight.beaveriot.rule.flow.builder;
+package com.milesight.beaveriot.rule.flow.yaml;
 
 import com.milesight.beaveriot.rule.RuleNodeInterceptor;
 import com.milesight.beaveriot.rule.flow.ComponentDefinitionCache;
@@ -8,8 +8,8 @@ import com.milesight.beaveriot.rule.model.flow.config.RuleChoiceConfig;
 import com.milesight.beaveriot.rule.model.flow.config.RuleConfig;
 import com.milesight.beaveriot.rule.model.flow.config.RuleFlowConfig;
 import com.milesight.beaveriot.rule.model.flow.config.RuleNodeConfig;
-import com.milesight.beaveriot.rule.model.flow.route.*;
-import com.milesight.beaveriot.rule.model.flow.route.base.OutputNode;
+import com.milesight.beaveriot.rule.model.flow.yaml.*;
+import com.milesight.beaveriot.rule.model.flow.yaml.base.OutputNode;
 import com.milesight.beaveriot.rule.support.RuleFlowIdGenerator;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.util.Assert;
